refactor: dissolve jjj folder, fix shortcut toggle colors, and resolve test failures
