+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import Menu from "@/app/Menu";
 import { MenuLayout } from "@/components/ui/Containers/Menu/MenuLayout";
-import { GameStateDashboard } from "@/jjj/GameStateDashboard";
+import { GameStateDashboard } from "@/components/ui/Layout/GameStateDashboard";
 
 export default function Home() {
 	return (
